fix: type price book

+++ NHbook-API/internal/models/cartItem.go
@@ -1,10 +1,10 @@
 package models
 
 type CartItem struct {
-	ID        uint    `json:"productID"`
-	Thumbnail string  `json:"thumbnail"`
-	Title     string  `json:"title"`
-	Price     float64 `json:"price"`
-	Quantity  int     `json:"quantity"`
-	Total     float64 `json:"total"`
+	ID        uint   `json:"productID"`
+	Thumbnail string `json:"thumbnail"`
+	Title     string `json:"title"`
+	Price     int    `json:"price"`
+	Quantity  int    `json:"quantity"`
+	Total     int    `json:"total"`
 }
